machine_platform_user_profile/requirements.txt
streamlit==1.25.0
mysql-connector-python==8.0.33
